requirements.txt
discord.py==1.6.0
python-dotenv==0.14.0